Missing cast

+++ src/map/skill.cpp
@@ -7684,7 +7684,7 @@ int skill_castend_nodamage_id (struct block_list *src, struct block_list *bl, ui
 	case SA_TAMINGMONSTER:
 		clif_skill_nodamage(src,*bl,skill_id,skill_lv);
 		if (sd != nullptr && dstmd != nullptr) {
-			std::vector<uint32> pet_list = { dstmd->mob_id };
+			std::vector<uint32> pet_list = { static_cast<uint32>(dstmd->mob_id) };
 			pet_catch_process1(*sd, pet_list);
 		}
 		break;
